+++ e2e/flows/import-from-bit-hub.e2e.js
@@ -54,6 +54,10 @@ describe('importing bit components from bitsrc.io', function () {
       const barFooPackageJson = helper.readPackageJson(path.join(helper.localScopePath, 'components', 'bar', 'foo'));
       expect(barFooPackageJson.name).to.equal('@bit/david.tests.bar.foo');
     });
+    it('should save the imported component as a dependency in the package.json of the project', () => {
+      const barFooPackageJson = helper.readPackageJson();
+      expect(barFooPackageJson.dependencies).to.deep.include({ '@bit/david.tests.bar.foo': './components/bar/foo' });
+    });
   });
   describe('when saveDependenciesAsComponents is set to TRUE in consumer bit.json', () => {
     before(() => {
@@ -148,6 +152,27 @@ describe('importing bit components from bitsrc.io', function () {
         const result = helper.runCmd('node app.js');
         expect(result.trim()).to.equal('got is-type v2 and got is-string');
       });
+    });
+  });
+  describe('installing as a package and then importing it', () => {
+    let packageJsonBeforeImport;
+    let packageJsonAfterImport;
+    before(() => {
+      helper.reInitLocalScope();
+      helper.runCmd('npm init -y');
+      helper.runCmd('npm i @bit/david.tests.utils.is-type --save');
+      packageJsonBeforeImport = helper.readPackageJson();
+      helper.runCmd('bit import david.tests/utils/is-type');
+      packageJsonAfterImport = helper.readPackageJson();
+    });
+    it('should not remove any property of the package.json created by npm', () => {
+      Object.keys(packageJsonBeforeImport).forEach(prop => expect(packageJsonAfterImport).to.have.property(prop));
+    });
+    it('should update the root package.json and change the dependency from a package to a local path', () => {
+      expect(packageJsonBeforeImport.dependencies['@bit/david.tests.utils.is-type']).to.equal('0.0.1');
+      expect(packageJsonAfterImport.dependencies['@bit/david.tests.utils.is-type']).to.equal(
+        './components/utils/is-type'
+      );
     });
   });
 });